+++ utils.c
@@ -6,17 +6,17 @@
 /*   By: aaizenbe <aaizenbe@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/18 14:32:33 by aaizenbe          #+#    #+#             */
-/*   Updated: 2024/03/18 14:32:33 by aaizenbe         ###   ########.fr       */
+/*   Updated: 2024/03/27 13:39:02 by aaizenbe         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "push_swap.h"
 
-int	ft_atoi(const char *str)
+long	ft_atoi(const char *str)
 {
-	int	i;
-	int	neg;
-	int	res;
+	int		i;
+	int		neg;
+	long	res;
 
 	i = 0;
 	neg = 1;
